Resolve GCI0010/GCI0012 hardcoded-credential duplicate findings
GCI0010 (Hardcoding and Configuration) and GCI0012 (Security Risk) both
matched the same credential-assignment patterns (apikey/password/secret/
token/api_key/pwd), producing two findings on the same line for every
hardcoded credential.

Fix: make GCI0012 the single owner of credential detection.

- GCI0012: replace crude full-line scan with the smarter LHS-only logic
  ported from GCI0010. Now skips comment lines, requires real assignment
  (not ==/!=/<=/>=), checks the secret keyword only on the variable-name
  side, skips lines whose only string literal is an ALL_CAPS env-var
  reference, and attaches FilePath + line number to the finding.
  'token' added to its pattern set.
- GCI0010: drop CheckSecrets + SecretNamePatterns + the
  HasAssignment / FindAssignmentIndex / IsEnvVarName helpers. Updated
  doc comment to make the new boundary explicit.
- Tests: migrated the type-name false-positive guards (HtmlTokenType in
  comparison, SecurityTokenInvalidException in nameof) from GCI0010Tests
  to GCI0012Tests, plus a new EnvVarNameOnly_ShouldNotFlag test.
- The CommentedCredential test in GCI0012Tests inverted (commented-out
  code is no longer treated as a hardcoded credential, consistent with
  the rest of the rule's file/comment handling).

All 978 tests pass. Verified end-to-end on the demo's
opts.ApiKey hardcoded literal: 1 finding (GCI0012) instead of 2.
