+++ packaging/GauntletCI.nuspec
@@ -122,6 +122,7 @@ v2.0.2
 - GitHub annotations now include Lines, Evidence, Why, and Action (multi-line via `%0A`) instead of just the one-line Summary.
 - PR review inline comments now use the run-log layout (RuleId — RuleName, Summary, Evidence list, Why, Action).
 - Diagnostic stderr logging on successful PR review and check-run posts so silent failures are easier to spot.
+- GCI0010 vs GCI0012 hardcoded-credential overlap resolved: GCI0012 (Security Risk) is now the single owner of credential-assignment detection. GCI0010 retains IPs, URLs, connection strings, ports, and environment names. GCI0012's heuristic was upgraded to LHS-only matching with comment + env-var-name skip (no more false positives on type names like `HtmlTokenType`).
 
 v2.0.1
 - New CLI flags: --github-pr-comments (post findings as PR review with inline comments) and --github-checks (post via GitHub Checks API)
