+++ src/GauntletCI.Tests/Rules/GCI0012Tests.cs
@@ -56,6 +56,45 @@ public class GCI0012Tests
     }
 
     [Fact]
+    public async Task SecretVariable_AssignedStringLiteral_ShouldFlag()
+    {
+        var diff = MakeDiff("    var myToken = \"placeholder-value\";");
+        var findings = await Rule.EvaluateAsync(diff, null);
+
+        Assert.Contains(findings, f => f.Summary.Contains("token"));
+    }
+
+    [Fact]
+    public async Task TypeNameContainingToken_InComparison_ShouldNotFlag()
+    {
+        // "token" appears in a type name (HtmlTokenType) in a comparison, not in a variable assignment.
+        var diff = MakeDiff("    if (_type == HtmlTokenType.StartTag) return;");
+        var findings = await Rule.EvaluateAsync(diff, null);
+
+        Assert.DoesNotContain(findings, f => f.Summary.Contains("token"));
+    }
+
+    [Fact]
+    public async Task TypeNameContainingToken_WithStringLiteralOnRhs_ShouldNotFlag()
+    {
+        // "token" appears in a type name on the right-hand side; the variable name itself is neutral.
+        var diff = MakeDiff("    var prefix = \"Microsoft.IdentityModel.\" + nameof(SecurityTokenInvalidException);");
+        var findings = await Rule.EvaluateAsync(diff, null);
+
+        Assert.DoesNotContain(findings, f => f.Summary.Contains("token"));
+    }
+
+    [Fact]
+    public async Task EnvVarNameOnly_ShouldNotFlag()
+    {
+        // String literal looks like an env var name (ALL_CAPS_UNDERSCORES) — this is a key reference, not a hardcoded value.
+        var diff = MakeDiff("    var apiKey = Environment.GetEnvironmentVariable(\"MY_API_KEY\");");
+        var findings = await Rule.EvaluateAsync(diff, null);
+
+        Assert.DoesNotContain(findings, f => f.Summary.Contains("hardcoded") || f.Summary.Contains("credential"));
+    }
+
+    [Fact]
     public async Task ParameterizedSql_ShouldNotFlag()
     {
         var diff = MakeDiff("    var sql = \"SELECT * FROM Users WHERE Id = @id\";");
@@ -74,13 +113,13 @@ public class GCI0012Tests
     }
 
     [Fact]
-    public async Task CommentedCredential_StillFlags()
+    public async Task CommentedCredential_DoesNotFlag()
     {
-        // Note: GCI0012 doesn't filter comments - this documents actual behavior
-        // (credentials in comments are still risky as they can end up in version control)
+        // GCI0012 skips comment lines for the credential check — commented-out code is
+        // historically common (TODOs/examples) and treating it as a hardcoded credential is noisy.
         var diff = MakeDiff("    // var apiKey = \"test1234\";");
         var findings = await Rule.EvaluateAsync(diff, null);
 
-        Assert.Contains(findings, f => f.Summary.Contains("hardcoded") || f.Summary.Contains("credential") || f.Summary.Contains("apikey"));
+        Assert.DoesNotContain(findings, f => f.Summary.Contains("hardcoded") || f.Summary.Contains("credential"));
     }
 }
